Renamed commands to better match DNN naming conventions
