+++ Modules/UpendoPrompt/Upendo.Modules.UpendoPrompt.csproj
@@ -77,6 +77,10 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Commands\Deprecated\DebugInfo.cs" />
+    <Compile Include="Commands\Deprecated\DebugMode.cs" />
+    <Compile Include="Commands\Deprecated\PopupMode.cs" />
+    <Compile Include="Commands\Deprecated\ThemesUsed.cs" />
     <Compile Include="Commands\ThemesUsed.cs" />
     <Compile Include="Commands\DebugInfo.cs" />
     <Compile Include="Commands\DebugMode.cs" />
