+++ Modules/UpendoPrompt/UpendoPrompt.dnn
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <dotnetnuke type="Package" version="6.0">
   <packages>
-    <package name="Upendo.Modules.UpendoPrompt" type="Module" version="01.03.00">
+    <package name="Upendo.Modules.UpendoPrompt" type="Module" version="01.03.01">
       <friendlyName>UpendoPrompt</friendlyName>
       <description>
         <![CDATA[<p>This extension adds some prompt commands that might help your management of DNN be easier from time to time. </p>]]>
@@ -52,7 +52,7 @@
             <attributes>
               <businessControllerClass>Upendo.Modules.UpendoPrompt.Components.FeatureController, Upendo.Modules.UpendoPrompt</businessControllerClass>
               <desktopModuleID>[DESKTOPMODULEID]</desktopModuleID>
-              <upgradeVersionsList>01.00.00,01.01.00,01.01.01,01.02.00,01.03.00</upgradeVersionsList>
+              <upgradeVersionsList>01.00.00,01.01.00,01.01.01,01.02.00,01.03.00,01.03.01</upgradeVersionsList>
             </attributes>
           </eventMessage>
         </component>
@@ -68,7 +68,7 @@
             <script type="UnInstall">
               <path>Providers\DataProviders\SqlDataProvider</path>
               <name>Uninstall.SqlDataProvider</name>
-              <version>01.03.00</version>
+              <version>01.03.01</version>
             </script>
           </scripts>
         </component>
@@ -78,7 +78,7 @@
             <assembly>
               <name>Upendo.Modules.UpendoPrompt.dll</name>
               <path>bin</path>
-              <version>01.03.00</version>
+              <version>01.03.01</version>
             </assembly>
           </assemblies>
         </component>
