+++ Modules/UpendoPrompt/UpendoPrompt_Symbols.dnn
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <dotnetnuke type="Package" version="6.0">
   <packages>
-    <package name="Upendo.Modules.UpendoPrompt_Symbols" type="Library" version="01.03.00">
+    <package name="Upendo.Modules.UpendoPrompt_Symbols" type="Library" version="01.03.01">
       <friendlyName>UpendoPrompt Module Symbols</friendlyName>
       <description><![CDATA[This extension adds some prompt commands that might help your management of DNN be easier from time to time. ]]></description>
       <owner>
@@ -14,7 +14,7 @@
       <releaseNotes src="ReleaseNotes.txt" />
       <azureCompatible>True</azureCompatible>
       <dependencies>
-        <dependency type="managedPackage" version="1.3.0">Upendo.Modules.UpendoPrompt</dependency>
+        <dependency type="managedPackage" version="1.3.1">Upendo.Modules.UpendoPrompt</dependency>
       </dependencies>
 
       <components>
